moved abstract object creation to Flow/factory and added array support
